Added check for classes that have mutually-incompatible base classes due to generic type argument mismatches. This addresses #5748. (#5762)

+++ packages/pyright-internal/src/tests/samples/classes7.py
@@ -10,7 +10,7 @@ class BaseClass(Generic[T]):
     pass
 
 
-IntBaseClass = BaseClass[int]
+IntBaseClass = BaseClass[float]
 
 
 # This should generate an error because the same
